Reconfigured Inputs for Game
I have gone through and remade new axes for all the games inputs, and
made axes for player 2. In game manager there is now a PlayerInput class
that is responsible for holding all the correct axes for a player. The
game manager holds two of the playerInput objects, one for p1 and 1 for
p2.  The parrot and Pirate get their inputsManager from the game manager
based on their player Number.