Implemented Phases + add Animator Controllers
I have started to implement phase switching for the game. It will run
each phase for the time  defined in the game manager. You will need to
use the Rough Test Camera system for right now till we add in the
splitscreen cameras. I have also gone and update pirate and parrot
prefabs. There is now one for each player. Also I have gone through and
made a new Animator for the Captain and the Parrot.