TImer now works, Fix treasure signal, & commented out setting manager
The timer and hud is now in and implemented in the main scene. The
treasure signal was having a weird error where it could be found when it
was put in the scene. but the moment you leave and return to the scene
it would not work. so I fixed it by making it a serialize field.

I had to comment out most of setting manager because it was not
implemented correctly and was throw errors and breaking stuff.