Added Respawn Button + cannons don't move anymore
I have added a killswitch to allow the player to choose to respawn at
the last island they reached. This will allow the player to easily
continue even if they get stuck. This button is map to the back(select)
button.

I also removed the rigidbody from the cannons so they can no longer be
moved.