Ground Cone Cast for Pirate
I have altered the pirates ground check to use a series of raycast that
cover a cone shape check. for each raycast that does not hit it will
shoot the next raycast in the chain. for example if down doesn't hit. it
will shoot the forwards raycast, then backwards, then left , then
finally the right raycast to see if the player is grounded. I also
increased the ground check distance to 0.3 and reduced the radius of the
pirates capsule collider to match 0.3.