FreeLook Mode added to parrot camera
The Parrot Camera now has 2 different modes that can be switch by
pressing R3. The first mode is flyMode in which the camera will always
stay behind the player and will be angled down. The player will be
unable to move the camera in this mode.

The second mode is FreeLook mode which allows the player to fully move
the camera as the fly. Note the camera will not stay behind parrot while
in the mode.