fixed Beartrap Sliding and Utility momentum issue.
I have removed momentum from utilities because it is too hard to aim
with right now. If we were going to keep it we would need a reticle and
we don't have time to implement that for imagine RIT.

I also fixed it so that the moment the beartrap hits the ground it will
freeze to the ground so that we don't have he shuffle board sliding
issue.